+++ src/Api/Filter/MatchFilter.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace App\Api\Filter;
+
+use ApiPlatform\Elasticsearch\Filter\AbstractFilter;
+use ApiPlatform\Metadata\Operation;
+use Symfony\Component\PropertyInfo\Type;
+
+/**
+ * This class represents a filter that performs a search based on matching properties in a given resource.
+ */
+final class MatchFilter extends AbstractFilter
+{
+    public function apply(array $clauseBody, string $resourceClass, Operation $operation = null, array $context = []): array
+    {
+        $properties = $this->getProperties($resourceClass);
+        $matches = [];
+
+        /** @var string $property */
+        foreach ($properties as $property) {
+            $matches[] = ['match' => [$property => $context['filters'][$property]]];
+        }
+
+        return isset($matches[1]) ? ['bool' => ['should' => $matches]] : $matches[0];
+    }
+
+    public function getDescription(string $resourceClass): array
+    {
+        if (!$this->properties) {
+            return [];
+        }
+
+        $description = [];
+        foreach ($this->properties as $filterParameterName => $value) {
+            $description[$filterParameterName] = [
+                'property' => $filterParameterName,
+                'type' => Type::BUILTIN_TYPE_STRING,
+                'required' => false,
+                'description' => 'Search field based on value given',
+                'openapi' => [
+                    'allowReserved' => false,
+                    'allowEmptyValue' => true,
+                    'explode' => false,
+                ],
+            ];
+        }
+
+        return $description;
+    }
+}
